AG-13178 - Move topology series base class to enterprise
